# server/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Partidazo API",
    "version": "1.0.0",
    "description": "API para la aplicación Partidazo - Plataforma de actividades deportivas",
    "contact": {
      "name": "Partidazo Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Servidor de desarrollo"
    }
  ],
...
  "paths": {
...
    "/users/{_id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Obtener un usuario por ID",
        "description": "Devuelve la información detallada de un usuario específico",
        "parameters": [
          {
            "in": "path",
            "name": "_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID único del usuario"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/users/add": {
      "post": {
        "tags": ["Users"],
        "summary": "Registrar un nuevo usuario",
        "description": "Crea una nueva cuenta de usuario en la plataforma",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "nombre", "apellido"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "usuario@ejemplo.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "Password123!"
                  },
                  "nombre": {
                    "type": "string",
                    "example": "Juan"
                  },
                  "apellido": {
                    "type": "string",
                    "example": "Pérez"
                  },
                  "imagenPerfil": {
                    "type": "string",
                    "description": "URL de la imagen de perfil"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Usuario creado exitosamente"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Email ya registrado o datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/users/follow": {
      "put": {
        "tags": ["Users"],
        "summary": "Seguir o dejar de seguir a un usuario",
        "description": "Actualiza la lista de seguidos del usuario actual y la lista de seguidores del usuario objetivo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["currentUser", "targetedUser"],
                "properties": {
                  "currentUser": {
                    "type": "object",
                    "description": "Información del usuario que realiza la acción"
                  },
                  "targetedUser": {
                    "type": "object",
                    "description": "Información del usuario a seguir/dejar de seguir"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Acción completada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/loggedin": {
      "get": {
        "tags": ["Auth"],
        "summary": "Obtener información del usuario autenticado",
        "description": "Devuelve la información del usuario actualmente logueado",
        "responses": {
          "200": {
            "description": "Usuario autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/loggedout/{email}": {
      "delete": {
        "tags": ["Auth"],
        "summary": "Cerrar sesión del usuario",
        "description": "Elimina la sesión actual del usuario",
        "parameters": [
          {
            "in": "path",
            "name": "email",
            "required": true,
            "schema": {
              "type": "string",
              "format": "email"
            },
            "description": "Email del usuario"
          }
        ],
        "responses": {
          "200": {
            "description": "Sesión cerrada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Usuario cerró sesión exitosamente"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/posts/creator/{_id}": {
      "get": {
        "tags": ["Posts"],
        "summary": "Obtener actividades creadas por un usuario",
        "description": "Devuelve todas las actividades publicadas por un usuario específico",
        "parameters": [
          {
            "in": "path",
            "name": "_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del usuario creador"
          }
        ],
        "responses": {
          "200": {
            "description": "Actividades del creador",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number"
                    },
                    "posts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Post"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/posts/joiner/{_id}": {
      "get": {
        "tags": ["Posts"],
        "summary": "Obtener actividades en las que el usuario participa",
        "description": "Devuelve todas las actividades a las que se ha unido un usuario",
        "parameters": [
          {
            "in": "path",
            "name": "_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del usuario participante"
          }
        ],
        "responses": {
          "200": {
            "description": "Actividades en las que participa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number"
                    },
                    "posts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Post"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/posts/{_id}": {
      "get": {
        "tags": ["Posts"],
        "summary": "Obtener una actividad por ID",
        "description": "Devuelve la información detallada de una actividad específica",
        "parameters": [
          {
            "in": "path",
            "name": "_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la actividad"
          }
        ],
        "responses": {
          "200": {
            "description": "Actividad encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "404": {
            "description": "Actividad no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/posts/add": {
      "post": {
        "tags": ["Posts"],
        "summary": "Publicar una nueva actividad",
        "description": "Crea una nueva actividad deportiva",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["creator_id", "titulo", "deporte", "fecha", "ubicacion"],
                "properties": {
                  "creator_id": {
                    "type": "string",
                    "example": "507f1f77bcf86cd799439011"
                  },
                  "titulo": {
                    "type": "string",
                    "example": "Partido de fútbol"
                  },
                  "deporte": {
                    "type": "string",
                    "example": "Fútbol",
                    "enum": ["Fútbol", "Baloncesto", "Tenis", "Pádel", "Running", "Ciclismo", "Natación", "Otro"]
                  },
                  "descripcion": {
                    "type": "string",
                    "example": "Partido amistoso de fútbol 7"
                  },
                  "fecha": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-12-15T18:00:00.000Z"
                  },
                  "ubicacion": {
                    "type": "object",
                    "properties": {
                      "lat": {
                        "type": "number",
                        "example": 37.3886
                      },
                      "lng": {
                        "type": "number",
                        "example": -5.9823
                      },
                      "direccion": {
                        "type": "string",
                        "example": "Polideportivo Norte, Sevilla"
                      }
                    }
                  },
                  "maxParticipantes": {
                    "type": "number",
                    "example": 14
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Actividad creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Actividad creada exitosamente"
                    },
                    "post": {
                      "$ref": "#/components/schemas/Post"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/posts/delete/{_id}": {
      "delete": {
        "tags": ["Posts"],
        "summary": "Eliminar una actividad",
        "description": "Elimina permanentemente una actividad del sistema",
        "parameters": [
          {
            "in": "path",
            "name": "_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la actividad a eliminar"
          }
        ],
        "responses": {
          "200": {
            "description": "Actividad eliminada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Post eliminado exitosamente"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/post/updateJoining": {
      "put": {
        "tags": ["Posts"],
        "summary": "Unirse o salirse de una actividad",
        "description": "Permite a un usuario unirse o retirarse de una actividad",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["postId", "userId"],
                "properties": {
                  "postId": {
                    "type": "string",
                    "description": "ID de la actividad"
                  },
                  "userId": {
                    "type": "string",
                    "description": "ID del usuario"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Acción completada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/post/cancelActivity/{postId}": {
      "delete": {
        "tags": ["Posts"],
        "summary": "Cancelar una actividad",
        "description": "Cancela una actividad (solo el creador puede hacerlo)",
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la actividad a cancelar"
          }
        ],
        "responses": {
          "200": {
            "description": "Actividad cancelada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Actividad cancelada"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/add-login-session": {
      "post": {
        "tags": ["Session"],
        "summary": "Crear sesión de usuario",
        "description": "Agrega datos del usuario a la sesión express cuando inicia sesión",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sesión creada exitosamente"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID único del usuario",
            "example": "507f1f77bcf86cd799439011"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email del usuario",
            "example": "usuario@ejemplo.com"
          },
          "password": {
            "type": "string",
            "description": "Contraseña hasheada del usuario"
          },
          "nombre": {
            "type": "string",
            "description": "Nombre del usuario",
            "example": "Juan"
          },
          "apellido": {
            "type": "string",
            "description": "Apellido del usuario",
            "example": "Pérez"
          },
          "imagenPerfil": {
            "type": "string",
            "description": "URL de la imagen de perfil"
          },
          "following": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string"
                }
              }
            },
            "description": "Lista de usuarios seguidos"
          },
          "followers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string"
                }
              }
            },
            "description": "Lista de seguidores"
          },
          "activitiesPosted": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string"
                }
              }
            },
            "description": "Actividades publicadas por el usuario"
          },
          "activitiesJoined": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string"
                }
              }
            },
            "description": "Actividades a las que se ha unido"
          },
          "notifications": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Notificaciones del usuario"
          }
        }
      },
      "Post": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID único del post/actividad",
            "example": "507f1f77bcf86cd799439012"
          },
          "creator_id": {
            "type": "string",
            "description": "ID del creador del post",
            "example": "507f1f77bcf86cd799439011"
          },
          "titulo": {
            "type": "string",
            "description": "Título de la actividad",
            "example": "Partido de fútbol 7"
          },
          "deporte": {
            "type": "string",
            "description": "Tipo de deporte",
            "example": "Fútbol",
            "enum": ["Fútbol", "Baloncesto", "Tenis", "Pádel", "Running", "Ciclismo", "Natación", "Otro"]
          },
          "descripcion": {
            "type": "string",
            "description": "Descripción de la actividad",
            "example": "Partido amistoso, todos los niveles bienvenidos"
          },
          "fecha": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha y hora de la actividad",
            "example": "2025-12-15T18:00:00.000Z"
          },
          "ubicacion": {
            "type": "object",
            "properties": {
              "lat": {
                "type": "number",
                "description": "Latitud",
                "example": 37.3886
              },
              "lng": {
                "type": "number",
                "description": "Longitud",
                "example": -5.9823
              },
              "direccion": {
                "type": "string",
                "description": "Dirección completa",
                "example": "Polideportivo Norte, Sevilla"
              }
            }
          },
          "participando": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "description": "ID del participante"
                },
                "nombre": {
                  "type": "string"
                },
                "imagenPerfil": {
                  "type": "string"
                }
              }
            },
            "description": "Lista de participantes"
          },
          "maxParticipantes": {
            "type": "number",
            "description": "Número máximo de participantes",
            "example": 14
          },
          "messages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "string"
                },
                "userName": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                },
                "timestamp": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            },
            "description": "Mensajes del chat de la actividad"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de creación del post"
          },
          "estado": {
            "type": "string",
            "description": "Estado de la actividad",
            "enum": ["activa", "cancelada", "completada"],
            "example": "activa"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Error en la solicitud"
          },
          "error": {
            "type": "string",
            "description": "Detalles adicionales del error"
          }
        }
      }
    }
  }
}